Changed auto to use a base class
